Completed exercises 4.17 - 4.23

+++ Part4/bloglist/controller/blogs.js
@@ -1,24 +1,52 @@
 const blogsRouter = require('express').Router()
-const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const {userExtractor} = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find()
+    const blogs = await Blog.find().populate('user', {username: 1, name: 1, id: 1})
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const body = request.body
 
-    if (!blog.title || !blog.url) {
+    const user = request.user
+    if (!user) {
+        return response.status(400).json({error: 'invalid userId'})
+    }
+
+    if (!body.title || !body.url) {
         return response.status(400).end()
     }
 
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    })
+
     const result = await blog.save()
+    user.blogs = user.blogs.concat(result)
+    await user.save()
     response.status(201).json(result)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({error: 'blog not found'})
+    }
+
+    if (blog.user.toString() !== user.id.toString()) {
+        return response.status(401).json({error: 'only the creator can delete this blog'})
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
